reviewed code for conditional rendering, yet to check functionality, can check once endpoints complete
