+++ Components/Profile/Profile.js
@@ -12,6 +12,38 @@ export class Profile extends Component {
     this.props.navigation.navigate("Home")
   }
 
+  renderClient = () => {
+    return <View>
+      <Text style={styles.userInfo}>Street Adress: {this.props.user.street_address}</Text>
+      <Text style={styles.userInfo}>City: {this.props.user.city}</Text>
+      <Text style={styles.userInfo}>State: {this.props.user.state}</Text>
+      <Text style={styles.userInfo}>Zip Code: {this.props.user.zip}</Text>
+      <View style={styles.infoCntainer}>
+        <Text style={styles.userInfoList}>Needs:</Text>
+        {allNeeds}
+      </View>
+      <View style={styles.infoCntainer}>
+        <Text style={styles.userInfoList}>Allergies: </Text>
+          {allAllergies}
+      </View>
+      <View style={styles.infoCntainer}>
+        <Text style={styles.userInfoList}>Dietary Restrictions: </Text>
+          {allRestrictions}
+      </View>
+      <View style={styles.infoCntainer}>
+        <Text style={styles.userInfoList}>Medications:</Text>
+        {allMedications}
+      </View>
+    </View>
+  }
+
+  renderCaretaker = () => {
+    return <View style={styles.infoCntainer}>
+      <Text style={styles.userInfoList}>Abilities:</Text>
+      {allAbilities}
+    </View>
+  }
+
   render() {
     let allNeeds = this.props.user.needs.map(need => {
       return <Text style={styles.userInfoList} key={Math.random()}>- {need}</Text>
@@ -24,6 +56,9 @@ export class Profile extends Component {
     })
     let allRestrictions = this.props.user.diet_restrictions.map(restr => {
       return <Text style={styles.userInfoList} key={Math.random()}>- {restr}</Text>
+    })
+    let allAbilities = this.props.user.diet_restrictions.map(ablility => {
+      return <Text style={styles.userInfoList} key={Math.random()}>- {ablility}</Text>
     })
     return (
       <View>
@@ -46,28 +81,10 @@ export class Profile extends Component {
         <ScrollView style={styles.profileContainer}>
           <Text style={styles.userInfo}>Username: {this.props.user.username}</Text>
         <Text style={styles.userInfo}>Name: {this.props.user.name}</Text>
-        <Text style={styles.userInfo}>Street Adress: {this.props.user.street_address}</Text>
-        <Text style={styles.userInfo}>City: {this.props.user.city}</Text>
-        <Text style={styles.userInfo}>State: {this.props.user.state}</Text>
-        <Text style={styles.userInfo}>Zip Code: {this.props.user.zip}</Text>
         <Text style={styles.userInfo}>Email: {this.props.user.email}</Text>
         <Text style={styles.userInfo}>Phone Number: {this.props.user.phone}</Text>
-        <View style={styles.infoCntainer}>
-          <Text style={styles.userInfoList}>Needs:</Text>
-           {allNeeds}
-        </View>
-        <View style={styles.infoCntainer}>
-          <Text style={styles.userInfoList}>Allergies: </Text>
-            {allAllergies}
-        </View>
-        <View style={styles.infoCntainer}>
-          <Text style={styles.userInfoList}>Dietary Restrictions: </Text>
-            {allRestrictions}
-        </View>
-        <View style={styles.infoCntainer}>
-          <Text style={styles.userInfoList}>Medications:</Text>
-           {allMedications}
-        </View>
+          {this.props.user.accountType === "client" && this.renderClient}
+            {this.props.user.accountType === "caretaker" && this.renderCaretaker}
         </ScrollView>
       </View>
     );
