+++ Components/Tasks/Tasks.js
@@ -192,7 +192,7 @@ export class Tasks extends Component {
               {task.due_date !== null && <Text style={styles.listItemSecond}>due: {task.due_date}</Text>}
               </View>
             )} 
-            {/* Need to add ability to complete functionality */}
+            {/* Need to add  un/complete functionality */}
           </View>
         </View>
     }).reverse();
@@ -205,9 +205,10 @@ export class Tasks extends Component {
     <View>
       <Text>{list.name}</Text>
       <Text>My Tasks</Text>
-        <Text>{!this.props.task? "No Tasks" : null}</Text>
-        <View>{this.props.user.accountType === 'client' ? this.clientTasks : this.caretakerTasks}</View> 
-        {/* insted of ternary use line 191 logic */}
+        <Text>{!this.props.task && "No Tasks"}</Text>
+        <View>
+        {this.props.user.accountType === 'client' && this.clientTasks}
+        {this.props.user.accountType === 'caretaker' && this.caretakerTasks}  </View> 
       </View>
     );
   }
